zcbor: Make changes to zcbor code to satisfy mynewt compile options
printf format type and bit-casting between uint and float.

+++ boot/zcbor/include/zcbor_print.h
@@ -73,17 +73,17 @@ do { \
 #endif
 
 __attribute__((used))
-static void zcbor_print_compare_lines(const uint8_t *str1, const uint8_t *str2, uint32_t size)
+static void zcbor_print_compare_lines(const uint8_t *str1, const uint8_t *str2, size_t size)
 {
-	for (uint32_t j = 0; j < size; j++) {
+	for (size_t j = 0; j < size; j++) {
 		zcbor_do_print("%x ", str1[j]);
 	}
 	zcbor_do_print("\r\n");
-	for (uint32_t j = 0; j < size; j++) {
+	for (size_t j = 0; j < size; j++) {
 		zcbor_do_print("%x ", str2[j]);
 	}
 	zcbor_do_print("\r\n");
-	for (uint32_t j = 0; j < size; j++) {
+	for (size_t j = 0; j < size; j++) {
 		zcbor_do_print("%x ", str1[j] != str2[j]);
 	}
 	zcbor_do_print("\r\n");
@@ -91,10 +91,10 @@ static void zcbor_print_compare_lines(const uint8_t *str1, const uint8_t *str2, 
 }
 
 __attribute__((used))
-static void zcbor_print_compare_strings(const uint8_t *str1, const uint8_t *str2, uint32_t size)
+static void zcbor_print_compare_strings(const uint8_t *str1, const uint8_t *str2, size_t size)
 {
-	for (uint32_t i = 0; i <= size / 16; i++) {
-		zcbor_do_print("line %d (char %d)\r\n", i, i*16);
+	for (size_t i = 0; i <= size / 16; i++) {
+		zcbor_do_print("line %zu (char %zu)\r\n", i, i*16);
 		zcbor_print_compare_lines(&str1[i*16], &str2[i*16],
 			MIN(16, (size - i*16)));
 	}
@@ -102,12 +102,12 @@ static void zcbor_print_compare_strings(const uint8_t *str1, const uint8_t *str2
 }
 
 __attribute__((used))
-static void zcbor_print_compare_strings_diff(const uint8_t *str1, const uint8_t *str2, uint32_t size)
+static void zcbor_print_compare_strings_diff(const uint8_t *str1, const uint8_t *str2, size_t size)
 {
 	bool printed = false;
-	for (uint32_t i = 0; i <= size / 16; i++) {
+	for (size_t i = 0; i <= size / 16; i++) {
 		if (memcmp(&str1[i*16], &str2[i*16], MIN(16, (size - i*16))) != 0) {
-			zcbor_do_print("line %d (char %d)\r\n", i, i*16);
+			zcbor_do_print("line %zu (char %zu)\r\n", i, i*16);
 			zcbor_print_compare_lines(&str1[i*16], &str2[i*16],
 				MIN(16, (size - i*16)));
 			printed = true;
